Lwip 通用文件梳理 (#2)
* [modify] extract the common files for different version about lwip.

* [update] the initialization logic of lwIP.
