+++ components/net/lwip/lwip-2.0.3/SConscript
@@ -26,7 +26,6 @@ src/core/tcp_out.c
 src/core/timeouts.c
 src/core/udp.c
 src/netif/ethernet.c
-src/netif/ethernetif.c
 src/netif/lowpan6.c
 """)
